+++ R/Option.R
@@ -11,3 +11,27 @@ Option <- S7::new_class(
   "Option",
   abstract = TRUE
 )
+
+#' @export
+Some <- S7::new_class(
+  "Some",
+  parent = Option,
+  properties = list(
+    value = S7::new_property(
+      class = S7::class_any,
+      setter = function(self, value) {
+        if (!is.null(self@value)) {
+          stop("@value is read-only")
+        }
+        self@value <- value
+        self
+      }
+    )
+  )
+)
+
+#' @export
+Nothing <- S7::new_class(
+  "Nothing",
+  parent = Option
+)