+++ NEWS.md
@@ -1,3 +1,4 @@
 # resultr (development version)
 
 * Initial CRAN submission.
+* Added `Result`, `Success` and `Failure` `S7` classes.